+++ test/new.jl
@@ -290,7 +290,7 @@ end
             @test_skip Pkg.test("Permutations")
         else
             Pkg.test("Permutations")
-        end 
+        end
     end
 end
 
@@ -2101,9 +2101,9 @@ end
         Pkg.add(url="https://github.com/00vareladavid/Unregistered.jl")
         Pkg.status(; io = io)
         @test occursin(r"Status `.+Project\.toml`", readline(io))
-        @test occursin(r"\[7876af07\] Example v\d\.\d\.\d `.+`", readline(io))
-        @test occursin(r"\[682c06a0\] JSON v0.18.0", readline(io))
-        @test occursin(r"\[dcb67f36\] Unregistered v\d\.\d\.\d `https://github\.com/00vareladavid/Unregistered\.jl#master`", readline(io))
+        @test occursin(r"\[7876af07\] Example\s*v\d\.\d\.\d\s*`.+`", readline(io))
+        @test occursin(r"\[682c06a0\] JSON\s*v0.18.0", readline(io))
+        @test occursin(r"\[dcb67f36\] Unregistered\s*v\d\.\d\.\d\s*`https://github\.com/00vareladavid/Unregistered\.jl#master`", readline(io))
         @test occursin(r"\[d6f4376e\] Markdown", readline(io))
     end
     ## status warns when package not installed
@@ -2113,15 +2113,17 @@ end
         io = PipeBuffer()
         Pkg.status(; io=io)
         @test occursin(r"Status `.+Project.toml`", readline(io))
-        @test occursin(r"→ \[7876af07\] Example v\d\.\d\.\d", readline(io))
+        @test occursin(r"→⌃ \[7876af07\] Example\s*v\d\.\d\.\d", readline(io))
         @test occursin(r"\[d6f4376e\] Markdown", readline(io))
-        @test "Info packages marked with → not downloaded, use `instantiate` to download" == strip(readline(io))
+        @test "Info Packages marked with → are not downloaded, use `instantiate` to download" == strip(readline(io))
+        @test "Info Packages marked with ⌃ have new versions available" == strip(readline(io))
         Pkg.status(;io=io, mode=Pkg.PKGMODE_MANIFEST)
         @test occursin(r"Status `.+Manifest.toml`", readline(io))
-        @test occursin(r"→ \[7876af07\] Example v\d\.\d\.\d", readline(io))
+        @test occursin(r"→⌃ \[7876af07\] Example\s*v\d\.\d\.\d", readline(io))
         @test occursin(r"\[2a0f44e3\] Base64", readline(io))
         @test occursin(r"\[d6f4376e\] Markdown", readline(io))
-        @test "Info packages marked with → not downloaded, use `instantiate` to download" == strip(readline(io))
+        @test "Info Packages marked with → are not downloaded, use `instantiate` to download" == strip(readline(io))
+        @test "Info Packages marked with ⌃ have new versions available" == strip(readline(io))
     end
     # Manifest Status API
     isolate(loaded_depot=true) do
@@ -2135,7 +2137,7 @@ end
         Pkg.add("Markdown")
         Pkg.status(; io=io, mode=Pkg.PKGMODE_MANIFEST)
         @test occursin(r"Status `.+Manifest.toml`", readline(io))
-        @test occursin(r"\[7876af07\] Example v0\.3\.0", readline(io))
+        @test occursin(r"\[7876af07\] Example\s*v0\.3\.0", readline(io))
         @test occursin(r"\[2a0f44e3\] Base64", readline(io))
         @test occursin(r"\[d6f4376e\] Markdown", readline(io))
     end
@@ -2170,14 +2172,16 @@ end
         ## diff project
         Pkg.status(; io=io, diff=true)
         @test occursin(r"Diff `.+Project\.toml`", readline(io))
-        @test occursin(r"\[7876af07\] \+ Example v0\.3\.0", readline(io))
+        @test occursin(r"\[7876af07\] \+ Example\s*v0\.3\.0", readline(io))
         @test occursin(r"\[d6f4376e\] - Markdown", readline(io))
+        @test occursin("Info Packages marked with ⌃ have new versions available", readline(io))
         ## diff manifest
         Pkg.status(; io=io, mode=Pkg.PKGMODE_MANIFEST, diff=true)
         @test occursin(r"Diff `.+Manifest.toml`", readline(io))
-        @test occursin(r"\[7876af07\] \+ Example v0\.3\.0", readline(io))
+        @test occursin(r"\[7876af07\] \+ Example\s*v0\.3\.0", readline(io))
         @test occursin(r"\[2a0f44e3\] - Base64", readline(io))
         @test occursin(r"\[d6f4376e\] - Markdown", readline(io))
+        @test occursin("Info Packages marked with ⌃ have new versions available", readline(io))
         ## diff project with filtering
         Pkg.status("Markdown"; io=io, diff=true)
         @test occursin(r"Diff `.+Project\.toml`", readline(io))
@@ -2202,7 +2206,7 @@ end
         Pkg.add(Pkg.PackageSpec(name="Example", version="0.4.0"); io=devnull)
         Pkg.status(; outdated=true, io=io)
         str = String(take!(io))
-        @test occursin("[7876af07] Example v0.4.0 (<v$v)", str)
+        @test occursin(Regex("⌃\\s*\\[7876af07\\] Example\\s*v0.4.0\\s*\\(<v$v\\)"), str)
         open(Base.active_project(), "a") do io
             write(io, """
                   [compat]
@@ -2211,7 +2215,7 @@ end
         end
         Pkg.status(; outdated=true, io=io)
         str = String(take!(io))
-        @test occursin("[7876af07] Example v0.4.0 [<v0.4.1], (<v$v)", str)
+        @test occursin(Regex("⌃\\s*\\[7876af07\\] Example\\s*v0.4.0\\s*\\[<v0.4.1\\], \\(<v$v\\)"), str)
     end
 end